Refactor `menu` to `pages` && Fix flicker when closing confirmation menu
