+++ src/main/java/io/redstudioragnarok/fbp/gui/pages/PageExperiments.java
@@ -1,7 +1,6 @@
-package io.redstudioragnarok.fbp.gui.menu;
+package io.redstudioragnarok.fbp.gui.pages;
 
 import io.redstudioragnarok.fbp.gui.BasePage;
-import io.redstudioragnarok.fbp.gui.GuiBase;
 import net.minecraft.client.gui.GuiButton;
 
 import java.io.IOException;
